Added Refresh Token, CORS headers
JWT token can be refreshed by sending GET request to hostname/auth/refresh-token. In 'Authorization' header 'refreshToken' needs to be send. Response JSON with 'token' and 'tokenTimeout'(in hour) will be sent.
Additionally, 'refreshToken' and 'refreshTokenTimeout'(in hour) is sent in cookie in resposne to successful login with 'token' and 'tokenTimeout' in response body.

CORS policy is set to all sites and http verbs. Headers allowed 'Content-Type', 'Authorization'